変更

+++ db/migrate/20220918040449_create_groups.rb
@@ -1,7 +1,7 @@
 class CreateGroups < ActiveRecord::Migration[6.1]
   def change
     create_table :groups do |t|
-      t.integer :owner_customer_id, null: false
+      t.bigint :owner_customer_id, null: false
       t.string  :name,              null: false
       t.text    :introduction,      null: false
       t.timestamps
